🐛(back) use a big int for video size field
The video model is using an IntegerField for the size field. The max
value available is 2147483647. As we are storing bytes, this represents
a little bit more than 2 gigbytes. This is not enough for us, we have to
use the BigIntegerField to store bigger number.

+++ src/backend/marsha/core/models/video.py
@@ -267,7 +267,7 @@ class Video(BaseFile, RetentionDateObjectMixin):
         null=True,
         blank=True,
     )
-    size = models.IntegerField(
+    size = models.BigIntegerField(
         help_text=_("Size of the video in bytes"),
         null=True,
         blank=True,
